IGNITE-15817 review fixes
