+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/util/PageUtils.java
@@ -17,6 +17,7 @@
 
 package org.apache.ignite.internal.pagememory.util;
 
+import java.nio.ByteBuffer;
 import org.apache.ignite.internal.util.GridUnsafe;
 
 /**
@@ -251,5 +252,39 @@ public class PageUtils {
         assert off >= 0;
 
         GridUnsafe.putLong(addr + off, v);
+    }
+
+    /**
+     * Copies memory from one buffer to another.
+     *
+     * @param src    Source.
+     * @param srcOff Source offset in bytes.
+     * @param dst    Destination.
+     * @param dstOff Destination offset in bytes.
+     * @param cnt    Bytes count to copy.
+     */
+    public static void copyMemory(ByteBuffer src, long srcOff, ByteBuffer dst, long dstOff, long cnt) {
+        byte[] srcArr = src.hasArray() ? src.array() : null;
+        byte[] dstArr = dst.hasArray() ? dst.array() : null;
+        long srcArrOff = src.hasArray() ? src.arrayOffset() + GridUnsafe.BYTE_ARR_OFF : 0;
+        long dstArrOff = dst.hasArray() ? dst.arrayOffset() + GridUnsafe.BYTE_ARR_OFF : 0;
+
+        long srcPtr = src.isDirect() ? GridUnsafe.bufferAddress(src) : 0;
+        long dstPtr = dst.isDirect() ? GridUnsafe.bufferAddress(dst) : 0;
+
+        GridUnsafe.copyMemory(srcArr, srcPtr + srcArrOff + srcOff, dstArr, dstPtr + dstArrOff + dstOff, cnt);
+    }
+
+    /**
+     * Copies memory from one address to another.
+     *
+     * @param srcAddr Source.
+     * @param srcOff  Source offset in bytes.
+     * @param dstAddr Destination.
+     * @param dstOff  Destination offset in bytes.
+     * @param cnt     Bytes count to copy.
+     */
+    public static void copyMemory(long srcAddr, long srcOff, long dstAddr, long dstOff, long cnt) {
+        GridUnsafe.copyMemory(null, srcAddr + srcOff, null, dstAddr + dstOff, cnt);
     }
 }
